Breaking change: use state code rather than "description"
For process.Status(), linux is the only platform where it's possible to
get state descriptions. This is specific to the linux /proc/<pid>/status
file.

I think it would be better to get the more platform-neutral state codes,
which are also present on FreeBSD, OpenBSD, darwin, etc.

The state codes are also better documented on "man proc" and "man ps"