Drop all docker logic from filled process, revert to upstream.
We're moving all the docker stuff into our own repo in order to use it
elsewhere and have better control on it. This also allows us to remove
Docker from the filled process and to revert the docker package changes
back to upstream.

Going forward we should not use gopsutil/docker