Add backwards compatible ConnectionStat config
This is a way to enable altering of ConnectionStat querying w/o
having to rip a new major version.

The ideal solution would likely have a `BareConnectionStat`
embedded in a `ConnectionStat`, which has a Uids field
because then the public method could naturally omit
unrequested data.

The querying of Uids is so expensive that it makes
the call hard to justify in cron like scenarios.

I considered having an environmental variable, but
that could also cause problems because a caller
may want to query without Uids often and rarely
make the full call.