process: Fix data race on darwin
Fixes #1900

`registerFuncs` mutates globals, and is called when any process instance
methods to access CPU time or memory are used.

If 2 separate process instances are used concurrently to read CPU/memory
this results in a data race.

Avoid globals completely by storing the loaded functions on a struct
alongside the `*common.Library`, which avoids the race.